Added some test facts for representing the board.
I'm going to try and see if I can write a rule that behaves as it should for
the simplest possible case. Given this board: '.BW', which I've represented
with hard coded facts, I want it to spit out X = 0. Once I've got that I
should be able to make it handle longer chains of discs as well.