Fixed arithmetic operations.
Ok. Arithmetic didn't work as I thought it would. I've fixed it by adding a
clause that computes the number I want to use before I use it. I have no
idea of how to make that computation in the clause itself, which would look
better (to me).

I've been reading about unification and equality in the Art of Prolog and at
wikibooks. It's going to take a while to get used to.