First attempt at matching longer chains of discs.
Next test case is to add another black disc to the board to still have it
identify 0 as a valid move. Using recursion thankfully reduced the number of
arithmetic operations in the rule as well. Guess I should have seen that ;-)

These rules still only match the '.BW' and not the '.BBW' but that should
only take another rchain called recursively.