Added rules for searching left, up and down.
This is a copy-paste job. I have no idea if this can be expressed in a more
compact manner or not. It probably can, I just can't think of it. Now there
are 4 legal_move rules, one for each direction, and a total of 8 chain rules.
One recursive and one terminating for each direction.

Unfortunately it only finds 3 out of the 4 solutions I expected it to.