Removed all of the test data.
I've run through a whole bunch of tests on my little 4x4 board. I'm not sure
how to script them at the moment and what the best way to do TDD in Prolog is
but suffice to say that I'm pretty confident that the rules I have will find
any and all legal moves.

Next up is to load the position facts from file rather than have them hard
coded in the definition.