Added test1.txt input file and a framework for reading files.
I've struggled quite a bit with Prolog's I/O functionality. I found this
snippet of code on http://en.wikibooks.org/wiki/Prolog/Input_and_Output and
I'll use it as a foundation for getting the board representation from the
text file. As is, process('filename') reads and prints the contents of a
file.