Added find_moves('filename') and print_moves to conclude this kata.
It appears as is if I'm done. Calling find_moves with either of the test
files prints out the correct results. I'll add some more tests later on and
try to figure out how to script these tests.

find_moves uses findall to populate Result with all the solutions and then
passes it on to print_moves which prints one at a time as R:C. Guess I could
spend some time to make it print out A1 instead of 1:1. Come to think of it.
I probably should.