Added the code neccessary to print A1 instead of 1:1.
That was actually easier than I thought it would be, once I found the
char_code function it took a while before I realized that I could use it *as
is* to go both ways. I did actually search for a code_char function ;-)
Guess I've got some way still to go.

Basic idea is to add 64 to the row which yields an ASCII code of 65-72 which
is translated into a Character and printed. Destructuring lists and elements
in lists is really simple. I was sure I was going to struggle with getting
at R and C from Move but my first guess turned out to work.