Added a separate file for a test framework.
I still haven't found anything about TDD in Prolog. There are some dead links
if you Google and a few papers mention it but I've not found anything I can
use. Earlier today I was reading about Higher Order Programming (HOP) in
Prolog and realized that I ought to be able to write a simple test framework
myself, in just a few lines of code. It's not going to be xUnit or anything
but it should be enough for my purposes.