Added the test rule.
test is basically find_moves but without the pretty printing stuff. By
providing Expected as input to the rule it will fail unless Expected can be
unified with the results from running findall on legal_move. If that ain't
the foundation of a test framework I don't know *what* is.