Added some output.
It's nice to see which files have been tested and what the result is. I read
that A -> B ; C is the Prolog way to define an if-then-else construct and it
seems to fit in quite nicely.

Next up I'll have to try and see what happens if one test fails.