Added a cut to prevent backtracking from reading a closed file.
Ok. After a short but intense debugging session using print(1) before every
get_char in the reversi.pl code it turns out that when findall fails, it
backtracks to some choice point that's in the middle of reading from the
input file. Which has been closed already. I added a cut just before the
findall and it takes care of the unexpected exception but it still doesn't
print failed, it just stops.