Added another rule in order to separate printing from running the test.
Adding another layer actually fixes the problem. But that was merely a side
effect. What I was really trying to do was to have it print out something
like: testing test1.txt fail! Expected Foo, got Bar. and so I needed a
Result variable but I don't want to have to specify that in test so I
separated running the tests into two parts.

I guess the reason this version works is because of how backtracking is
performed. Perhaps it is limited to the context it is defined which would
mean that once it failed run(Filename, Expected) it had exhausted all
choice points when it came back up to test.