Moved unification of Result and Expected to test.
I'd like to have this into a generic utility, if possible. So it seems to me
that run ought to be focused on running the test to produce a Result and that
the test should worry about whether or not that's a fail or pass and print
whatever is appropriate.