Switched places in the if-then-else clause to fix printout problem.
Ok. Once again, I was bitten by backtracking. Switching places in the if-then
else clause fixed the problem with print(Result) printing _560 instead of the
actual result. I guess it makes sense if I think about how the -> must be
implemented to work. Once it goes into the ; (else) part it would pass a
choice point and clear any variables.

Also, I changed all print to write since it turns out that print is gprolog
specific. Well, it's at least not in the ISO standard. I'll fix the reversi
file as well.