Broke the tight coupling between test and run.
Up until now, test could only work with run. In order for it to be more
general it needs to be completely ignorant about what code it needs to call.
The only assumption still built into it is that the last argument is Result.
I can live with that for the moment.

You can build new functors dynamically with =.. and run them with call. It's
undoubtably a powerful concept but I think that it feels a bit clunky.
Perhaps I haven't quite grokked it yet but it feels slightly limited. I doubt
that this code would work if run required 3 or more arguments.