Made test take the comparison predicate as an argument instead.
I moved out the comparison predicate of Expected and Result to the arg list.
The predicate is built dynamically using =.. which means I can define a
whole bunch of test functions. I also renamed those to assertXX in xUnit
style. I think the text that's printed must also be changed but all in all
I'm quite pleased with how generic the test method is becoming.