Added eunit tests.
Erlang, of course, includes a unit test framework. I suspected as much but
didn't find one at first glance so I decided to start writing code instead.
This morning I googled a bit more and realized that eunit was actually
included in the distro I've got setup. So I've added the test cases I've been
running manually to test my functions.

My use of eunit is not the most elegant. I've added all tests to the reversi
module. I hope to put them in a separate one eventually but this will have to
do for now.