Added function skeletons for reading a file and constructing the board.
The plan is to read the text file, row by row and add them to Rows. Once
that's done I plan to use a list comprehension (I think) to create a Cols
list. When I have the board represented in this way, it should be fairly easy
to apply the two regexps to get at the legal moves.

Note! The matching function doesn't correctly handle rows like: '.BW.BW..'.