Added the code neccessary to read in the file.
The differences between Prolog and Erlang are beginning to show. Only one
definition of the same function name/arity combination. Means I have to use
case to split up functionality. Also, return values FTW! I've been,
completely unneccessary, passing in return variables in the calls. Glad to
see that Erlang works this way, it was one of the big minuses of Prolog IMHO.
Third, I love that erlang has namespaces. And fourth, batteries included! So
happy to see that there's plenty of libraries available to do stuff. Also, I
was a bit surprised to see that my distro came with eunit, which is not part
of the language. I like this little language more and more. LFE might turn
out a winner.