Added implementation of make_cols.
This was trickier than I thought it'd be. I've been trying to wrap my head
around list comprehensions but I gave up and took the easy recursive solution
instead. I'm not sure I like all of these 'end's that keeping having to be
inserted in 'case's, 'if's and 'fun's. I sure hope LFE (Lisp Flavoured
Erlang) is better in this respect. I suspect the next round of commits will
mostly be refactorings since this code is ugly.