Fixed find_l/rmoves so they correctly handle several matches.
This took a while actually. I struggled with a looping construct to extract
positions from the match data. Don't even know why I bothered with that. It
seems that recursion with an accumulator is *the* way to go.

Also thought I could use just one extract function but I split it into two
since they're different enough in how they interpret the match data.

Feels as if I'm finally getting into the groove of Erlang. It's definitely a
step in the right direction from Prolog and it'll be interesting to try out
some concurrent programming.