Done. Turned on the acceptance tests and implemented regex matching for views.
Picking up the views from each position and checks if any view matches
the regex. If so, that is a legal move.

Excercise done. I liked the regex approach, but since fetching strings in
all directions was done by tokenizing and building strings, perhaps the
regex didn't add that much value? In a recursive solution, the views
would likely become unnessecary and the checks could be done in place
instead. Will try that solution as well for comparison.