Moved the tests to reversi.pl.
In an effort to make test.pl as re-usable as possible I need to remove
anything specific to the reversi code. That means moving all and run. It's
still a bit ugly in the sense that you have to pass in a Result variable that
pretty much takes over how you write the test.