Removed all module specifiers to calls in reversi module.
Ok. Removing the export of all but one of the functions worked. I also had to
remove the module specifier in the calls but now that that's been done I can
honestly say I like the code better. I'm all for namespaces but I'm going to
find out if I can type append instead of lists:append everywhere. That would
kind of make me happy.