Added a failing test for diagonals.
At the meetup last week we discussed algorithms and ways of solving this
kata. I had misunderstood some of the requirements for being done. I have not
made a solution that finds all legal moves. It doesn't find any of the
diagonal moves. I suspect this will be quite tricky to get working with the
code as it looks at the moment but I'll give it a try.