Added extract_moves_diagonals.
Including diagonal moves in the solution set is probably going to take a lot
of re-organizing of the code. There's not going to be an easy way of
extracting positions from matches in the Diagonals array. Instead of focusing
on Rows, Cols and Diagonals (as I do now) I think I'll arrange the code
around lmoves and rmoves and pass in the starting position of the string to
match against and a direction to calculate all other positions. That way, we
should be able to get rid of a lot of code.