Added a failing test for the new matching function.
As I said in the previous commit, I suspect much of the code I've written so
far can be deleted. Instead of passing in Rows, Cols and Diagonals I'm
thinking I can pass in a bunch of strings grouped by direction as in {1,0}
means each char in the string add 1 to the initial column position. So
instead of Rows I could pass in [{1,0},[[{1,1},........], ...]] where the
inner array contains a position identifier and the string representing the
row. When I find a match in one of these strings I pull out the position that
matched (say 4 for example) and multiply that with the direction (giving
{4,0} in this case) and add it to the starting position ({1,1}) which gives
us {5,1} for this example.

The above ought to be able to handle both columns, rows and diagonals in the
same way. Perhaps it'll all end up as one map operation. Oh the joy!