Refactoring of code to make it work with the new match function.
Ok, here goes. This is a monster commit. But it's mostly copy-paste stuff.
There were several places that needed some attention. First of all the
load_game_state/1 function should return the whole board as one list (nested)
containing Rows, Cols and Diagonals. That, in turn, meant that I had to
update the tests for it. It was a lot of grunt work but now that it's in
place we can move on to the more interesting task of modifying the find_moves
to use the new match function. That was pretty easy actually and not very
interesting. I just updated the struct being matched as the result from
load_game_state and passed the whole thing on to the new matcher. This also
allowed me to remove the extract_moves_XXX functions that was previously
used. Phew. Ok, all done.

After having removed a lot of tests we end up with one failing test. I'll
investigate what's up with that next.