Added another test + input file.
I added test3.txt and a test for it. This makes the test suite on par with
the Prolog solution. Not that I would want to say that the Prolog code is
particularly well tested but still. Apart from the fact that the Prolog
solution uses R,C instead of C,R as Erlang does they match perfectly. What
a relief\!