Added a test for findMove.
I realize that things came in the wrong order there. I'm still trying to get
my head wrapped around how to do things in Haskell. For the moment I plan to
use simple test functions returning the and'ed result of several test
expressions but once I figure out how to use QuickCheck I'll move on to that
instead.

The plan is to do this similar to how I did the Prolog version. I'll try to
match the simplest possible case to begin with and not worry too much about
IO and such.