First stab at making the test pass.
Evaluating test_findMove still returns False but there are a few things that
have happened. First of all, I've modified the function signature to return a
list of Ints instead and secondly I've added a function called findMoveInRow
which takes care of the searching. At the moment it returns the position of
the first player char it encounters.

I suspect I'll do another function to search for and check that positions
found actually are valid moves. Plenty of things undecided and that I need to
learn.

Still not comfy enough with QuickCheck to put it to use. Make quirky work-
around will have to do a while longer.