Added a (failing) test for findMoveInRow.
At the moment, this is a fairly redundant test since it tests the exact same
thing as the other one, but I really want to have tests for each function so.

Next up is to try and write that function that checks whether this is indeed
a valid move or not.