Added an implementation of the findChain function.
Ok, so this is basically very simple. If we find a '.' we return False. If we
find something that's not a '.' and not the player char we return whatever
the recursive call returns. If we find a player char we return True if pos is
larger than 0.

I suspect this can be done prettier but it'll have to do for now. I can
*understand* it like this and that does a lot for me. Running test_findChain
returns True now BTW. So onto the next failing test.