Added findChain function definition for handling empty lists.
That's better. No more Program errors. Still a failing test though but that's
because findChain only searches in one direction. Essentially we end up with
the same problem as with the other solutions. Not sure how I'll want to
represent the board just yet. Perhaps I should rename this from matrix to
something else.