Modified the findChain function to take an increment parameter.
In order to make this work both ways (searching left and right) but without
having to define (basically) the same function two times I want this function
to take an increment parameter for starters and a function that traverses the
list as another parameter. This rewrite changes nothing in terms of making
tests pass. Everything that's been done is modifiying function calls to match
the new implementation.

Next up is making findChain take a function as a parameter that will be used
for traversing the list. Tail is hard coded into the function definition at
the moment.