Removed dropEmpty function and used Maybe instead.
Learning about Maybe and Just put my current solution in a slightly different
light. Instead of returning -1 for a no-match it now returns Nothing which it
seems is the Haskell way of doing things. So far so good.

Turns out though that Maybe/Just is not particularly intuitive. At least not
to me. Well. Perhaps it's just me being tired but it does feel slightly
cumbersome.