Started working on finding all moves in a row.
Since there can be more than one valid move per row, it is neccessary to make
findMoveInRow return a list of values. This actually also solves the problem
with NOT returning a value at all. In this instance I even think it works
better than using Maybe.

Unfortunately modifying findMoveInRow to return a list has the effect that
findMoves won't work as is. I've commented that out for now and will start to
reimplement it in a way that works with the new findMoveInRow function.