Added searching right in rows to complete the refactoring.
Ok. All tests pass. That completes the implementation of finding all moves in
a row and the refactoring I thought neccessary to have it done. I'm still not
happy with the reverseIndex and (reverse row) bits. I wonder if this could
be captured in a different (more elegant) manner.

Oh the joy of programming in a language that makes you feel like a butcher