Added implementation for valid moves in columns.
This implementation messed things up a bit in my opinion. Not happy at all
with how it turned out. Also, I noted that reverseIndex relies on the board
being 8x8 so I modified the test case to a real size board instead of mucking
about with reverseIndex implementation.

I also learned that Haskell is not fond of newlines. It complained like hell
when I tried to format my matrix with each row on a separate line. It said
something about unexpected commas, perhaps they can/should be left out. It is
after all WS-sensitive.