Added extraction methods for player and board + a test_all.
I still haven't come around to the actual IO. I know that readFile dumps a
file's contents into a String so I thought it'd be a good idea to have a way
to parse that string first. Ok. Excuses! Excuses! I know. It's time to tackle
the infamous Haskell IO capabilities.