Added a non-working process function!
This code reflects how I'd hope to be able to handle File IO. Unfortunately
Haskell isn't simple as that (and probably for good reason). If I've
understood correctly, tainted code is tainted forever. So the return value
that is wrapped in the IO monad is not calculated/evaluated or whatever until
forced to. One way of doing this seems to be using do blocks and actions, but
I'd rather keep the unsafe stuff as limited as possible. I was really hoping
to be able to use converters to get out of this mess. Guess not