Fixed (well, avoided might be a better description) IO problem.
Ok. Once tainted always tainted. Haskell IO might actually be understood as
an implementation of the religious concept 'original sin'.

I must admit that I found this 'solution' using a kind of brute force
approach. It may very well be that it can be rewritten in a much prettier
form. I'm just glad it doesn't span hundreds of SLOCs.