Rewrote all of my tests using HUnit.
Of course there's a HUnit! I was so certain that I was going to use
QuickCheck that I delayed writing my tests for a framework deliberately
(QuickCheck looks uber-complicated to me). Earlier, trying to read up on QC,
I learned that there is a much simpler testing framework called HUnit.
Excellent! It IS however a bit cumbersome to work with but what the hell. It
sure beats writing specific test functions returning True/False.