Refactored cols to use \!\! instead of my own function.
Ok. So I started coding before finsihing either of the tutorials I've been
reading. Yesterday I learned about \!\! which is the built-in function to
access an element in a list based on position. This makes my implementation
of nth redundant and the code probably becomes both faster and safer and
prettier as well.