Refactored findMoves to use transpose instead of cols.
Peter Marks (@peter_marks) suggested I use transpose instead of my own home-
brewed version of it (in cols). It sure simplifies my code but unfortunately
I'm going to need to generate diagonals at some point as well so I suspect
I'll be writing a more general transformer that can handle all 4 ways of
slicing and dicing that is neccessary to search through the whole board for
solutions.

Or, it'll turn out that Haskell indeed has something built-in but I'll
probably not know about until it's too late any way ;-)