Added placeholders for diagonal moves.
I'm hoping to solve the diagonal moves the same way as I've done the
horizontal and vertical ones. Trickiest part is writing the transformation
functions (dl and dr). Most of everything else is just boiler plate.

Before starting this I just learned about Template Haskell which is an add on
to the GHC (Glasgow Haskell Compiler) which allows for compile time meta
programming (macros). Perhaps, one day, I'll be able to rewrite my
findMovesX functions in that. Some day.