Added implementation for dr.
Ok, so now I'm done with both dl and dr. That gives me about everything I
need to find all diagonal moves as well. I can't say I'm pleased about how
the code turned out. But it works. Still not sure if my chasing that perfect
line of beautiful abstraction and generality is a hopeless business or not.

To quote the Beautiful South: 'I may be ugly, but I've got the bottle opener.'
As long as the code works...