Added implementation for findMovesDL.
findMovesDL and DR are trickier than C and R since both X and Y coordinates
change depending on where in the row a match is found. I'm reusing the
makeSlice function to get at the positions. Note that I've added an
implementation for reversed rows as well, even though I don't have a test
case for it. I'll go fix that now.