Added implementation for findMovesDR.
This was mostly a copy and paste job. It makes all tests pass. Just to be
sure I'll try to add some tricky ones tonight. But if all turns out well,
I'll consider this implementation done and will start working on a write up
of my experiences, problems, insights (yes, I've had some of those as well)
and a-ha moments.