Added more test files.
Seems that this implementation works as it should. It could use a little
refactoring though. Been thinking about if I should make a function for
generating rows and columns as well. It might make a few of the functions
simpler even though it adds redundant functionality (id and transpose).

I'll be tinkering with this for a while I suppose.